move highlighting method to util
